Remove CachedSource
That doesn't really serve much purpose (you don't want to cache all
reads) and can lead to bugs because it tries to read full page which may
be after EOF - leading to issues when using e.g. S3.

It seems we should instead cache either in each source or just when
reading the IFD